+++ src/services/API/index.ts
@@ -1,2 +1,0 @@
-export * from "./connect.js";
-export * from "./item.js";