+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.4.0] - 2024-01-30
+ 
+### Changed
+- Fix login and favorites of favorites retrieval
+- Update dependencies
+- Add discord notification
+
 ## [1.3.0] - 2023-07-10
  
 ### Changed